+++ tasks/task-html2tpl.js
@@ -24,7 +24,7 @@ module.exports = function(cfg,SILLY){
         filename:filename,
         basename:Path.basename(abs,'.js'),
         extname:Path.extname(abs,'.js')
-      },SILLY.var)
+      },{self:SILLY.config})
     commontask.read(filename)
     .then(function(buffer){
       var filecontent = buffer.toString()